requirements.txt
# ============================================================
# E-MAT - ETHICAL Malware Analysis Toolkit
# Unified Requirements File
# ============================================================
# Sections: [core], [desktop], [web]
# The emat.py launcher reads these sections and installs
# only the packages needed for the chosen interface.
# CLI uses [core] only. Desktop adds [desktop]. Web adds [web].
# ============================================================

# [core] - Required for all interfaces (CLI, Desktop, Web)
pefile>=2023.2.7
yara-python>=4.3.1
pyelftools>=0.31
oletools>=0.60.1
pdfminer.six>=20221105
capstone>=5.0.1
python-magic-bin>=0.4.14;sys_platform=="win32"
python-magic>=0.4.27;sys_platform!="win32"
docker>=7.0.0
scikit-learn>=1.3.0
requests>=2.31.0
colorama>=0.4.6
tabulate>=0.9.0
click>=8.1.7
Jinja2>=3.1.2

# [desktop] - Additional packages for PyQt6 desktop GUI
PyQt6>=6.6.0

# [web] - Additional packages for Flask web interface
Flask>=3.0.0
Flask-CORS>=4.0.0
